+++ example/lib/slides/slides.dart
@@ -1,3 +1,4 @@
+export 'background_slide.dart';
 export 'blank_slide.dart';
 export 'code_highlight_slide.dart';
 export 'drawer_slide.dart';