+++ example/lib/slides/steps_slide.dart
@@ -12,9 +12,6 @@ class StepsSlide extends FlutterDeckSplitSlide {
         );
 
   @override
-  Color? get rightBackgroundColor => Colors.white10;
-
-  @override
   Widget left(BuildContext context) {
     return FlutterDeckBulletList(
       useSteps: true,
@@ -35,7 +32,10 @@ class StepsSlide extends FlutterDeckSplitSlide {
         'widget. This way, you can reveal or hide content, run animations, '
         'etc.\n\nFlutterDeckBulletList widget (the one on the left) supports '
         'steps out of the box.',
-        style: Theme.of(context).textTheme.titleLarge,
+        style: Theme.of(context)
+            .textTheme
+            .titleLarge
+            ?.copyWith(color: Theme.of(context).colorScheme.onPrimary),
       ),
     );
   }