+++ lib/src/templates/blank_slide.dart
@@ -59,7 +59,4 @@ abstract class FlutterDeckBlankSlide extends FlutterDeckSlideBase {
           )
         : null;
   }
-
-  @override
-  Widget? background(BuildContext context) => null;
 }