+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # NEXT
 
 - feat: add FlutterDeckCodeHighlight widget
+- feat: make the background configurable for the whole slide deck
 - fix: drawer toggle
+- docs: update the example app and README.md
 
 # 0.1.0+3